Started working on magical effects
Today was kinda like black Friday here in China so I spent most of the
day on Taobao buying stuff, didn’t have much time to code so just
started working on the skeleton for the magic effect selection plus
minor changes in naming and how the weapon is displayed. Stuff is
coming together.